@@ -1,34 +1,32 @@
-name: budget_master
+name: budget_master # Or your actual app name
 description: "A new Flutter project."
 publish_to: 'none' 
 version: 1.0.0+1
 
-# -------------------- ADD THIS BLOCK --------------------
 environment:
-  sdk: '>=3.0.0 <4.0.0' # This is a robust way to define it
-# --------------------------------------------------------
+  sdk: '>=3.9.0 <4.0.0'
 
 dependencies:
   flutter:
     sdk: flutter
 
-  # Database
-  isar: ^3.1.0+1
-  isar_flutter_libs: ^3.1.0+1
-  path_provider: ^2.0.11
+  # --- CORE DEPENDENCIES ---
+  # Use the latest versions from pub.dev
+  objectbox: ^4.0.0
+  objectbox_flutter_libs: any 
+  flutter_riverpod: ^2.5.1
+  path_provider: ^2.1.3
+  intl: ^0.19.0
 
-  # State Management & DI
-  flutter_riverpod: ^2.3.6
-  
-  # Other useful packages
-  intl: ^0.18.1
-  
 dev_dependencies:
+  analyzer: ^8.2.0
   flutter_test:
     sdk: flutter
-    
-  build_runner: ^2.4.6
-  isar_generator: ^3.1.0+1
+  
+  # --- CORE DEV DEPENDENCIES ---
+  build_runner: ^2.4.9
+  objectbox_generator: ^4.0.0
+  flutter_lints: ^3.0.0
 
 flutter:
   uses-material-design: true
