@@ -1,15 +1,15 @@
-import 'package:isar/isar.dart';
+import 'package:objectbox/objectbox.dart';
 
-part 'account.g.dart'; // This will be generated by build_runner
-
-@collection
+@Entity() // Changed from @collection
 class Account {
-  Id id = Isar.autoIncrement; // You can also use id = null to auto increment
+  @Id() // Changed from Id
+  int id = 0;
 
-  @Index(type: IndexType.value, unique: true)
-  late String name; // e.g., "M-PESA", "CRDB", "Cash"
+  @Unique() // Changed from @Index(unique: true)
+  late String name;
 
   late double balance;
 
+  // No-args constructor is required by ObjectBox
   Account();
 }
