@@ -1,23 +1,16 @@
 import 'package:flutter_riverpod/flutter_riverpod.dart';
-import 'package:isar/isar.dart';
+import 'package:objectbox/objectbox.dart';
 import 'package:path_provider/path_provider.dart';
-import 'package:budget_master/domain/models/account.dart';
-import 'package:budget_master/domain/models/category.dart';
-import 'package:budget_master/domain/models/transaction.dart';
+import 'package:path/path.dart' as p;
+import 'package:budget_master/objectbox.g.dart'; 
 
-// A provider that asynchronously initializes Isar
-final isarProvider = FutureProvider<Isar>((ref) async {
+// A provider that asynchronously initializes ObjectBox and provides the Store
+final objectboxProvider = FutureProvider<Store>((ref) async {
   final dir = await getApplicationDocumentsDirectory();
-  if (Isar.instanceNames.isEmpty) {
-    return await Isar.open(
-      [
-        AccountSchema,
-        CategorySchema,
-        TransactionSchema,
-      ],
-      directory: dir.path,
-      inspector: true,
-    );
-  }
-  return Future.value(Isar.getInstance());
+
+  // The 'openStore' function is created by the code generator
+  final store =
+      await openStore(directory: p.join(dir.path, "obx-budget-master"));
+
+  return store;
 });
