@@ -6,10 +6,10 @@
 
 #include "generated_plugin_registrant.h"
 
-#include <isar_flutter_libs/isar_flutter_libs_plugin.h>
+#include <objectbox_flutter_libs/objectbox_flutter_libs_plugin.h>
 
 void fl_register_plugins(FlPluginRegistry* registry) {
-  g_autoptr(FlPluginRegistrar) isar_flutter_libs_registrar =
-      fl_plugin_registry_get_registrar_for_plugin(registry, "IsarFlutterLibsPlugin");
-  isar_flutter_libs_plugin_register_with_registrar(isar_flutter_libs_registrar);
+  g_autoptr(FlPluginRegistrar) objectbox_flutter_libs_registrar =
+      fl_plugin_registry_get_registrar_for_plugin(registry, "ObjectboxFlutterLibsPlugin");
+  objectbox_flutter_libs_plugin_register_with_registrar(objectbox_flutter_libs_registrar);
 }
