@@ -6,9 +6,9 @@
 
 #include "generated_plugin_registrant.h"
 
-#include <isar_flutter_libs/isar_flutter_libs_plugin.h>
+#include <objectbox_flutter_libs/objectbox_flutter_libs_plugin.h>
 
 void RegisterPlugins(flutter::PluginRegistry* registry) {
-  IsarFlutterLibsPluginRegisterWithRegistrar(
-      registry->GetRegistrarForPlugin("IsarFlutterLibsPlugin"));
+  ObjectboxFlutterLibsPluginRegisterWithRegistrar(
+      registry->GetRegistrarForPlugin("ObjectboxFlutterLibsPlugin"));
 }
